Updated file with comments

+++ factorial.c
@@ -1,8 +1,9 @@
 #include <stdio.h>
+#include <inttypes.h>
 
 int main () {
-	long long int factorial = 1;
-	int index = 1;
+	uint64_t factorial = 1;
+	uint8_t index = 1;
 	
 	printf ("Enter a number whose factorial you wish to check: ");
 	scanf ("%d", &index);
@@ -15,7 +16,9 @@ int main () {
 		printf ("Sorry you entered an index too high for this computer to compute a factorial of");
 	}
 	
-	int i;
+	//checks to make sure that input from user is valid and within boundaries of 
+	
+	uint8_t i;
 	
 	for (i=1; i<= index; i++) {
 		factorial = factorial * i;
